Added API R and Python tutorials
