+++ CHANGE_LOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 _________________________________________________________________________
+## 2023-05-06
+> ### Added 
+> - the API tutorials from Bitbucket
+> ### Updated 
+> - the LICENSE typo
+
 
 ## 2023-03-29
   
@@ -10,20 +16,17 @@ _________________________________________________________________________
 > - CONTRIBUTING.md
 > - CODE_OF_CONDUCT.md
 > - LICENSE.md
-> ### Changed
-> ### Fixed
+
 
 ## 2023-03-07
   
 > ### Added
 > - tutorial on how to work with AppEEARS output CSV file in the cloud
-> ### Changed
-> ### Fixed
+
 
 ## 2023-02-16
   
 > ### Added
 > - tutorials on how to work with AppEEARS COG outputs in the cloud
-> ### Changed
-> ### Fixed
+
