+++ Data/.netrc
@@ -1,0 +1,3 @@
+machine urs.earthdata.nasa.gov 
+login Your_USERNAME
+password Your_PASSWORD
