+++ Data/DixieFire.geojson
@@ -1,0 +1,1 @@
+{"type":"FeatureCollection","features":[{"type":"Feature","properties":{},"geometry":{"coordinates":[[[-120.13867467647607,39.495927433194424],[-120.13867467647607,41.21001513065187],[-121.79728009443659,41.21001513065187],[-121.79728009443659,39.495927433194424],[-120.13867467647607,39.495927433194424]]],"type":"Polygon"}}]}
