+++ Data/R_output/NPS-Vegetation-request.json
@@ -1,0 +1,1 @@
+{"tier": 2, "error": null, "params": {"dates": [{"endDate": "12-31-2018", "startDate": "01-01-2018"}], "layers": [{"layer": "LST_Day_1km", "product": "MOD11A2.061"}, {"layer": "LST_Night_1km", "product": "MOD11A2.061"}, {"layer": "Lai_500m", "product": "MCD15A3H.006"}], "coordinates": [{"id": "0", "category": "Grand Canyon", "latitude": 36.2062, "longitude": -112.1271}, {"id": "1", "category": "Zion", "latitude": 37.2893, "longitude": -112.9738}]}, "status": "processing", "created": "2023-05-06T21:32:03.022204", "task_id": "052d0826-cb7f-412f-963a-4b150dd97ab0", "updated": "2023-05-06T21:32:42.660094", "user_id": "mjami@contractor.usgs.gov", "estimate": {"request_size": 272, "request_memory": 1}, "retry_at": null, "has_swath": false, "task_name": "NPS Vegetation", "task_type": "point", "api_version": "v1", "svc_version": "3.28", "web_version": null, "has_nsidc_daac": false, "expires_on": "2023-06-05T21:32:42.660094", "attempts": 1}
