+++ Data/nps_boundary.prj
@@ -1,0 +1,1 @@
+GEOGCS["North_American_Datum_1983",DATUM["D_North_American_1983",SPHEROID["Geodetic Reference System of 1980",6378137.0,298.2572221009113]],PRIMEM["Greenwich",0.0],UNIT["Degree",0.0174532925199433]]
