+++ Data/point-request.json
@@ -1,0 +1,1 @@
+{"tier": 2, "error": null, "params": {"dates": [{"endDate": "03-01-2023", "recurring": false, "startDate": "01-01-2015", "yearRange": [1950, 2050]}], "layers": [{"layer": "LST_Day_1km", "product": "MOD11A1.061"}, {"layer": "LST_Night_1km", "product": "MOD11A1.061"}], "coordinates": [{"id": "EROS", "latitude": 43.736038, "longitude": -96.62139}]}, "status": "processing", "created": "2023-05-06T17:53:44.623072", "task_id": "7819e6c4-368c-474e-b4f8-56399ec00d2f", "updated": "2023-05-06T17:55:26.628644", "user_id": "mjami@contractor.usgs.gov", "estimate": {"request_size": 2982, "request_memory": 1}, "retry_at": null, "has_swath": false, "task_name": "EROS", "task_type": "point", "api_version": null, "svc_version": "3.28", "web_version": "3.28", "has_nsidc_daac": false, "expires_on": "2023-06-05T17:55:26.628644", "attempts": 1}
