+++ Python/setup.md
@@ -1,0 +1,42 @@
+# Python resources Setup Instructions
+
+
+ 
+
+## 1. Prerequisites:
+
+------------
+
+## 2. Procedures:  
+
+
+--------------------
+
+## 3. Netrc file configuration:
+
++   Setting up a netrc file:
+
+
+    
+------------------------------------------------------------------------
+
+
+## 4. R Environment Setup          
+
++   **Required packages:**
+
+```
+
+
+
+---
+
+## Contact Info:  
+
+Email: LPDAAC@usgs.gov  
+Voice: +1-866-573-3222  
+Organization: Land Processes Distributed Active Archive Center (LP DAAC)¹  
+Website: <https://lpdaac.usgs.gov/>  
+Date last modified: 05-06-2023  
+
+¹Work performed under USGS contract G15PD00467 for NASA contract NNG14HH33I.
