+++ README.md
@@ -19,24 +19,41 @@ Welcome to the AppEEARS-Data-Resources repository. This repository provides reso
 
 ## Repository Contents
 
-> Tutorial are stored in `scripts` folder and are listed below:
-
-> + **COG_AppEEARS_S3_Direct_Access.ipynb**: Jupyter Notebook demonstrating how to use AppEEARS Cloud Optimized GEOTIFF (COG) outputs using Python.
-> + **COG_AppEEARS_S3_Direct_Access.html**: HTML export of Jupyter Notebook demonstrating how to use AppEEARS Cloud Optimized GEOTIFF (COG) outputs using Python.
-> + **Point_Sample_AppEEARS_S3_Direct_Access.ipynb**: Jupyter Notebook demonstrating how to access AppEEARS point sample Comma-Separated Values (CSV) outputs using Python.
-> + **Point_Sample_AppEEARS_S3_Direct_Access.html**: HTML export of Jupyter Notebook demonstrating how to access AppEEARS point sample Comma-Separated Values (CSV) outputs using Python.
+> The supporting files for use cases are stored in `Data` folder.  
 
 
-> Defined functions that will be called in the scripts are store in `modules` folder.
+| Repository Folders | Summary | 
+|----|-----|
+| `**Data**` | Supporting files for use cases | 
+| `**Python**` | Python resources including tutorials, how-tos, scripts, Defined modules that will be called from the Python resources, and setup instructions | 
+| `**R**` |  R resources including tutorials, how-tos, scripts, Defined modules that will be called from the R resources, and setup instructions | 
 
-> The supporting files for use cases are stored in `additional_files` folder.
+> Python and R resources stored in this repositories are listed below:  
+
+
+| Repository Contents | Summary | Path |
+|----|-----|----|
+| **COG_AppEEARS_S3_Direct_Access.ipynb** | Jupyter Notebook demonstrating how to use AppEEARS Cloud Optimized GEOTIFF (COG) outputs using Python | `Python\tutorials` |
+|**COG_AppEEARS_S3_Direct_Access.html** | HTML export of Jupyter Notebook demonstrating how to use AppEEARS Cloud Optimized GEOTIFF (COG) outputs using Python | `Python\tutorials` |
+| **Point_Sample_AppEEARS_S3_Direct_Access.ipynb** | Jupyter Notebook demonstrating how to access AppEEARS point sample Comma-Separated Values (CSV) outputs using Python | `Python\tutorials` |
+| **Point_Sample_AppEEARS_S3_Direct_Access.html** | HTML export of Jupyter Notebook demonstrating how to access AppEEARS point sample Comma-Separated Values (CSV) outputs using Python | `Python\tutorials` |
+| **AppEEARS_API_Area.ipynb:** | Jupyter Notebook demonstrating how to use Python to connect to the AppEEARS API to submit and downlaod an area sample  | `Python\tutorials` |
+| **AppEEARS_API_Area.html:** | HTML demonstrating how to use Python to connect to the AppEEARS API to submit and downlaod an area sample  | `Python\tutorials` |
+| **AppEEARS_API_Point.ipynb:** | Jupyter Notebook demonstrating how to use Python to connect to the AppEEARS API to submit and downlaod a point sample | `Python\tutorials` |
+| **AppEEARS_API_Point.html:** | HTML demonstrating how to use Python to connect to the AppEEARS API to submit and downlaod a point sample | `Python\tutorials` |
+| |  | `R\tutorials` |
+| |  | `R\tutorials` |
+| |  | `R\tutorials` |
+
 ---
 
 ## Helpful Links    
+
 + [AppEEARS Website](https://appeears.earthdatacloud.nasa.gov/)
 + [AppEEARS Documentation](https://appeears.earthdatacloud.nasa.gov/help)
 + [AppEEARS API Documentation](https://appeears.earthdatacloud.nasa.gov/api/)
 + [LP DAAC Website](https://lpdaac.usgs.gov/)
++ [LP DAAC GitHub](https://github.com/nasa/LPDAAC-Data-Resources)
 
 
 ---
@@ -47,6 +64,6 @@ Email: LPDAAC@usgs.gov
 Voice: +1-866-573-3222  
 Organization: Land Processes Distributed Active Archive Center (LP DAAC)¹  
 Website: <https://lpdaac.usgs.gov/>  
-Date last modified: 02-06-2023  
+Date last modified: 05-06-2023  
 
 ¹Work performed under USGS contract G15PD00467 for NASA contract NNG14HH33I.
