--- additional_files/Dixie-Fire-request.json
@@ -1,1 +1,0 @@
-{"tier": 2, "error": null, "params": {"geo": {"type": "FeatureCollection", "features": [{"id": "0", "type": "Feature", "geometry": {"type": "Polygon", "coordinates": [[[-120.13867467647607, 39.495927433194424], [-120.13867467647607, 41.21001513065187], [-121.79728009443659, 41.21001513065187], [-121.79728009443659, 39.495927433194424], [-120.13867467647607, 39.495927433194424]]]}, "properties": {}}]}, "dates": [{"endDate": "07-31-2021", "startDate": "07-01-2021"}], "layers": [{"layer": "_500m_16_days_EVI", "product": "MOD13A1.061"}], "output": {"format": {"type": "geotiff"}, "projection": "geographic"}}, "status": "processing", "created": "2023-02-07T07:22:28.872424", "task_id": "8f9858b1-d9d1-4aef-a343-1a82e964f160", "updated": "2023-02-07T07:46:11.825562", "user_id": "mjami@contractor.usgs.gov", "estimate": {"request_size": 1867801.226138162, "request_memory": 1}, "retry_at": null, "has_swath": false, "task_name": "Dixie Fire", "task_type": "area", "api_version": "v1", "svc_version": "3.23", "web_version": null, "has_nsidc_daac": false, "expires_on": "2023-03-09T07:46:11.825562", "attempts": 1}
